Modularise the ISH commands. This allows the built in command set to be extended. The envbirmonent stuff really needs to be sorted!

+++ source/main.d
@@ -71,7 +71,7 @@ void sireArgs(ref string[] args)
 {
 	  
    // Read and variable definitions
-   int i = 0; // input parmeter
+   size_t i = 0; // input parmeter
    while (i < args.length)
    {
 	  string arg = args[i];
